Move src/components into app
