+++ src/app/feed.xml/route.ts
@@ -1,4 +1,4 @@
-import { makeRssResponse } from '@/src/libs/rss'
+import { makeRssResponse } from '../libs/rss'
 
 export async function GET() {
   return makeRssResponse()
