+++ src/app/music/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next'
 
-import Discography from '../../components/Discography'
-import { HugeHeading } from '../../components/HugeHeading'
+import Discography from '../components/Discography'
+import { HugeHeading } from '../components/HugeHeading'
 
 export const metadata: Metadata = {
   title: 'Music',
