+++ todo.md
@@ -1,6 +1,5 @@
 - back button in blog, where should it take you
 
 - test rss
-- sitemap generation
 - meta data
 - tracking
